Move Verifications query to decidim-verifications
