+++ decidim-verifications/app/controllers/decidim/verifications/authorizations_controller.rb
@@ -76,15 +76,15 @@ module Decidim
       end
 
       def active_authorization_methods
-        Decidim::Authorizations.new(user: current_user).pluck(:name)
+        Authorizations.new(user: current_user).pluck(:name)
       end
 
       def granted_authorizations
-        Decidim::Authorizations.new(user: current_user, granted: true)
+        Authorizations.new(user: current_user, granted: true)
       end
 
       def pending_authorizations
-        Decidim::Authorizations.new(user: current_user, granted: false)
+        Authorizations.new(user: current_user, granted: false)
       end
     end
   end
