+++ decidim-verifications/spec/commmands/decidim/verifications/authorize_user_spec.rb
@@ -15,7 +15,7 @@ module Decidim::Verifications
       )
     end
 
-    let(:authorizations) { Decidim::Authorizations.new(user: user, granted: true) }
+    let(:authorizations) { Authorizations.new(user: user, granted: true) }
 
     context "when the form is not authorized" do
       before do
